engine/controlnet/requirements.txt
torch>=2.0.0
diffusers>=0.18.0
transformers>=4.30.0
pillow>=9.0.0
numpy>=1.24.0
opencv-python>=4.7.0
trimesh>=3.9.0
accelerate>=0.20.0
xformers>=0.0.20
safetensors>=0.3.1 